Initial upload
First draft of notebook with working visualizations of TF-IDF text embeddings using T-SNE and KNN and some performance comparisons. Still requires further revisions and update for full functionality and testing of additional features. Future plans to add steps for a more complete study of the price match in progress. Data not included due to competition rules.